--- src/main/java/com/javafx/mavenproject/morfologicalTransfLibrary/Median.java
@@ -1,122 +1,0 @@
-package com.javafx.mavenproject.morfologicalTransfLibrary;
-
-import java.awt.*;
-import java.awt.image.BufferedImage;
-import java.lang.reflect.Array;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Map;
-
-public class Median {
-
-    public static BufferedImage medianFilter(BufferedImage bufferedImage, int windowWidth, int windowHeight) {
-        BufferedImage result = new BufferedImage(bufferedImage.getWidth(), bufferedImage.getHeight(), BufferedImage.TYPE_INT_RGB);
-
-        int width = bufferedImage.getWidth();
-        int height = bufferedImage.getHeight();
-
-        int[] window = new int[windowHeight * windowWidth];
-        int[][] image = ImageUtils.convertTo2DArray(bufferedImage);
-
-        int edgeX = windowWidth / 2;
-        int edgeY = windowHeight / 2;
-
-        Color[] pixel = new Color[9];
-
-        int[] A=new int[windowWidth*windowHeight];
-        int[] R=new int[windowWidth*windowHeight];
-        int[] B=new int[windowWidth*windowHeight];
-        int[] G=new int[windowWidth*windowHeight];
-
-        for(int i=edgeX; i<width-edgeX-1; i++) {
-            for(int j=edgeY; j<height-edgeY-1; j++) {
-                //int i=0;
-                int k = 0;
-                for(int fx=0; fx<windowWidth;fx++) {
-                    for(int fy=0; fy<windowHeight; fy++) {
-                        //window[i] = image[y+fy-edgeY][x+fx-edgeX];
-                        //i++;
-                /*pixel[0]=new Color(bufferedImage.getRGB(i-1,j-1));
-                pixel[1]=new Color(bufferedImage.getRGB(i-1,j));
-                pixel[2]=new Color(bufferedImage.getRGB(i-1,j+1));
-                pixel[3]=new Color(bufferedImage.getRGB(i,j+1));
-                pixel[4]=new Color(bufferedImage.getRGB(i+1,j+1));
-                pixel[5]=new Color(bufferedImage.getRGB(i+1,j));
-                pixel[6]=new Color(bufferedImage.getRGB(i+1,j-1));
-                pixel[7]=new Color(bufferedImage.getRGB(i,j-1));
-                pixel[8]=new Color(bufferedImage.getRGB(i,j));
-                for(int k=0;k<9;k++){
-                    R[k]=pixel[k].getRed();
-                    B[k]=pixel[k].getBlue();
-                    G[k]=pixel[k].getGreen();
-                }*/
-
-                A[k] = bufferedImage.getRGB(i + fx-edgeX, j+fy-edgeY) >> 24 & 0xff;
-                R[k] = bufferedImage.getRGB(i + fx-edgeX, j+fy-edgeY) >> 16 & 0xff;
-                G[k] = bufferedImage.getRGB(i + fx-edgeX, j+fy-edgeY) >> 8 & 0xff;
-                B[k] = bufferedImage.getRGB(i + fx-edgeX, j+fy-edgeY) & 0xff;
-                k++;
-            }
-                    //Arrays.sort(window);
-                    //result.setRGB(x, y, window[(windowWidth * windowHeight)/2]);
-                    //Arrays.sort(R);
-                    //Arrays.sort(G);
-                    //Arrays.sort(B);
-                    //double dist = calculateMinDistance(R, G, B);
-
-                    result.setRGB(i,j, calculateMinDistance(R, G, B));//convertFromRGBtoInt(R[4], G[4], B[4]));
-                }
-            }
-        }
-        return result;
-    }
-
-    private static int convertFromRGBtoInt(int R, int G, int B) {
-        int rgb = R;
-        rgb = (rgb << 8) + G;
-        rgb = (rgb << 8) + B;
-
-        return rgb;
-    }
-
-    //private static double calculateMinDistance(int R0, int G0, int B0, int R, int G, int B) {
-    private static int calculateMinDistance(int[] R, int[] G, int[] B) {
-        //double distance = Math.sqrt(Math.pow(R0-R, R0-R)+Math.pow(G0-G, G0-G) + Math.pow(B0-B, B0-B));
-        //double[] distance = new double[R.length];
-        double distanceMin = 0.0;
-        int minDistR = 0;
-        int minDistG = 0;
-        int minDistB = 0;
-        for(int i = 0; i<R.length; i++) {
-            double localMinDistance = 0.0;
-            for (int j = 0; j < R.length; j++) {
-                //if (i == R.length/2)
-                //continue;
-                //double distance = Math.sqrt(Math.pow(R[R.length/2]-R[i], 2)+Math.pow(G[R.length/2]-G[i], 2) + Math.pow(B[R.length/2]-B[i], 2));
-                //double distance = Math.sqrt(Math.pow(minDistR-R[i], 2) + Math.pow(minDistG-G[i], 2) + Math.pow(minDistB-B[i], 2));
-                //sumanyczna odległość w stosunku do inncyh wektorów
-                localMinDistance += Math.sqrt(Math.pow(R[i]-R[j], 2) + Math.pow(G[i]-G[j], 2) + Math.pow(B[i]-B[j], 2));
-
-                /*if ((distance < distanceMin) || (i == 0)) {
-                    distanceMin = distance;
-                    minDistR = R[i];
-                    minDistG = G[i];
-                    minDistB = B[i];
-                }*/
-                //map.put(distance[i], new int[])
-            }
-            if ((localMinDistance < distanceMin) || (i == 0)) {
-                distanceMin = localMinDistance;
-                minDistR = R[i];
-                minDistG = G[i];
-                minDistB = B[i];
-            }
-        }
-        //Arrays.sort(distance);
-        //System.out.println(distance[0]);
-
-        //return distance[1];
-        return convertFromRGBtoInt(minDistR, minDistG, minDistB);
-    }
-
-}
